Detect if SPOP service not running
If Spotify is not configured, SPOP will not be running.  This code will attempt to connect to SPOP several times and if it fails will no longer poll the SPOP service until restarted.  Note that this still causes a delay at start-up as the program attempts to connect to SPOP.