Fixed issues from my walkthrough test
Just finished a walkthrough of the commands for the Runeaudio script and made some changes as a result.  Mainly minor tweaks however I did remove the instructions for reinstalling the easy_setup system for python as pip2 appears to be functioning properly post install.  I'm not sure why it was not working in my previous tests.